+++ flash/image/backbones.py
@@ -39,23 +39,6 @@ if _TORCHVISION_AVAILABLE:
         "barlow-twins": HTTPS_VISSL + "barlow_twins/barlow_twins_32gpus_4node_imagenet1k_1000ep_resnet50.torch",
     }
 
-    def _fn_mobilenet_vgg(model_name: str, pretrained: bool = True) -> Tuple[nn.Module, int]:
-        model: nn.Module = getattr(torchvision.models, model_name, None)(pretrained)
-        backbone = model.features
-        num_features = 512 if model_name in VGG_MODELS else model.classifier[-1].in_features
-        return backbone, num_features
-
-    for model_name in MOBILENET_MODELS + VGG_MODELS:
-        _type = "mobilenet" if model_name in MOBILENET_MODELS else "vgg"
-
-        IMAGE_CLASSIFIER_BACKBONES(
-            fn=catch_url_error(partial(_fn_mobilenet_vgg, model_name)),
-            name=model_name,
-            namespace="vision",
-            package="torchvision",
-            type=_type
-        )
-
     def _fn_resnet(model_name: str,
                    pretrained: Union[bool, str] = True,
                    weights_paths: dict = {"supervised": None}) -> Tuple[nn.Module, int]:
@@ -95,14 +78,7 @@ if _TORCHVISION_AVAILABLE:
 
         return backbone, num_features
 
-    def _fn_resnet_fpn(
-        model_name: str,
-        pretrained: bool = True,
-        trainable_layers: bool = True,
-        **kwargs,
-    ) -> Tuple[nn.Module, int]:
-        backbone = resnet_fpn_backbone(model_name, pretrained=pretrained, trainable_layers=trainable_layers, **kwargs)
-        return backbone, 256
+
 
     for model_name in RESNET_MODELS:
         clf_kwargs = dict(
@@ -126,48 +102,11 @@ if _TORCHVISION_AVAILABLE:
             )
         IMAGE_CLASSIFIER_BACKBONES(**clf_kwargs)
 
-        OBJ_DETECTION_BACKBONES(
-            fn=catch_url_error(partial(_fn_resnet_fpn, model_name)),
-            name=model_name,
-            package="torchvision",
-            type="resnet-fpn"
-        )
 
-    def _fn_densenet(model_name: str, pretrained: bool = True) -> Tuple[nn.Module, int]:
-        model: nn.Module = getattr(torchvision.models, model_name, None)(pretrained)
-        backbone = nn.Sequential(*model.features, nn.ReLU(inplace=True))
-        num_features = model.classifier.in_features
-        return backbone, num_features
 
-    for model_name in DENSENET_MODELS:
-        IMAGE_CLASSIFIER_BACKBONES(
-            fn=catch_url_error(partial(_fn_densenet, model_name)),
-            name=model_name,
-            namespace="vision",
-            package="torchvision",
-            type="densenet"
-        )
 
-if _TIMM_AVAILABLE:
 
-    def _fn_timm(
-        model_name: str,
-        pretrained: bool = True,
-        num_classes: int = 0,
-        **kwargs,
-    ) -> Tuple[nn.Module, int]:
-        backbone = timm.create_model(model_name, pretrained=pretrained, num_classes=num_classes, **kwargs)
-        num_features = backbone.num_features
-        return backbone, num_features
 
-    for model_name in timm.list_models():
-
-        if model_name in TORCHVISION_MODELS:
-            continue
-
-        IMAGE_CLASSIFIER_BACKBONES(
-            fn=catch_url_error(partial(_fn_timm, model_name)), name=model_name, namespace="vision", package="timm"
-        )
 
 
 # Wide resnet models from self-supervised training pipelines
