+++ flash/image/backbones/resnet.py
@@ -1,6 +1,21 @@
-"""
-ResNet encoder adapted from: https://github.com/facebookresearch/swav/blob/master/src/resnet50.py
-"""
+# Copyright The PyTorch Lightning team.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+#
+# ResNet encoder adapted from: https://github.com/facebookresearch/swav/blob/master/src/resnet50.py
+# as the official torchvision implementation does not support wide resnet architecture
+# found in self-supervised learning model weights
 import torch
 import torch.nn as nn
